Add files via upload
The following code files were used for my thesis:
Thesis 1, 2, and 3 were used for the preprocessing and EDA steps, in order. 
RFE.IMB was for conducting the experiments on the Imbalanced dataset.
RUS was for conducting Random Undersampling and its experiments.
SMOGNN was for conducting SMOGN and its experiments.
significance_test was for conducting the paired t-test at the end.